+++ crypto/keys/keyring_test.go
@@ -96,8 +96,6 @@ func TestLazyKeyManagementKeyRing(t *testing.T) {
 
 	// addr cache gets nuked - and test skip flag
 	require.NoError(t, kb.Delete(n2, "", true))
-
-	require.NotPanics(t, kb.CloseDB)
 }
 
 // TestSignVerify does some detailed checks on how we sign and validate
@@ -157,7 +155,8 @@ func TestLazySignVerifyKeyRing(t *testing.T) {
 	// Import a public key
 	armor, err := kb.ExportPubKey(n2)
 	require.Nil(t, err)
-	kb.ImportPubKey(n3, armor)
+	err = kb.ImportPubKey(n3, armor)
+	require.NoError(t, err)
 	i3, err := kb.Get(n3)
 	require.NoError(t, err)
 	require.Equal(t, i3.GetName(), n3)
@@ -210,7 +209,8 @@ func TestLazySignVerifyKeyRing(t *testing.T) {
 
 	// Now try to sign data with a secret-less key
 	_, _, err = kb.Sign(n3, p3, d3)
-	require.NotNil(t, err)
+	require.Error(t, err)
+	require.Equal(t, "cannot sign with offline keys", err.Error())
 }
 
 func TestLazyExportImportKeyRing(t *testing.T) {
@@ -408,6 +408,6 @@ func TestSupportedAlgos(t *testing.T) {
 	t.Cleanup(cleanup)
 	kb, err := NewKeyring("keybasename", "test", dir, nil)
 	require.NoError(t, err)
-	require.Equal(t, []SigningAlgo([]SigningAlgo{"secp256k1"}), kb.SupportedAlgos())
-	require.Equal(t, []SigningAlgo([]SigningAlgo{"secp256k1"}), kb.SupportedAlgosLedger())
+	require.Equal(t, []SigningAlgo{"secp256k1"}, kb.SupportedAlgos())
+	require.Equal(t, []SigningAlgo{"secp256k1"}, kb.SupportedAlgosLedger())
 }