+++ types/rest/rest_test.go
@@ -235,8 +235,9 @@ func TestReadRESTReq(t *testing.T) {
 
 	// test OK
 	ReadRESTReq(w, req, codec.New(), &br)
-	require.Equal(t, BaseReq{ChainID: "alessio", Memo: "text"}, br)
 	res := w.Result()
+	t.Cleanup(func() { res.Body.Close() })
+	require.Equal(t, BaseReq{ChainID: "alessio", Memo: "text"}, br)
 	require.Equal(t, http.StatusOK, res.StatusCode)
 
 	// test non valid JSON
@@ -247,6 +248,7 @@ func TestReadRESTReq(t *testing.T) {
 	ReadRESTReq(w, req, codec.New(), &br)
 	require.Equal(t, br, br)
 	res = w.Result()
+	t.Cleanup(func() { res.Body.Close() })
 	require.Equal(t, http.StatusBadRequest, res.StatusCode)
 }
 
@@ -255,6 +257,7 @@ func TestWriteSimulationResponse(t *testing.T) {
 	w := httptest.NewRecorder()
 	WriteSimulationResponse(w, codec.New(), 10)
 	res := w.Result()
+	t.Cleanup(func() { res.Body.Close() })
 	require.Equal(t, http.StatusOK, res.StatusCode)
 	bs, err := ioutil.ReadAll(res.Body)
 	require.NoError(t, err)
@@ -361,13 +364,13 @@ func TestPostProcessResponseBare(t *testing.T) {
 	got, err = ioutil.ReadAll(res.Body)
 	require.NoError(t, err)
 	t.Cleanup(func() { res.Body.Close() })
-	require.Equal(t, []string([]string{"application/json"}), res.Header["Content-Type"])
+	require.Equal(t, []string{"application/json"}, res.Header["Content-Type"])
 	require.Equal(t, `{"error":"couldn't marshal"}`, string(got))
 }
 
 type badJSONMarshaller struct{}
 
-func (_ badJSONMarshaller) MarshalJSON() ([]byte, error) {
+func (badJSONMarshaller) MarshalJSON() ([]byte, error) {
 	return nil, errors.New("couldn't marshal")
 }