+++ x/bank/genesis.go
@@ -16,7 +16,9 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, genState GenesisState) {
 			panic(err)
 		}
 
-		keeper.SetBalances(ctx, balance.Address, balance.Coins)
+		if err := keeper.SetBalances(ctx, balance.Address, balance.Coins); err != nil {
+			panic(fmt.Errorf("error on setting balances %w", err))
+		}
 	}
 }